finished

+++ libft/ft_atoi.c
@@ -14,8 +14,8 @@
 
 int	ft_atoi(const char *str)
 {
-	int				sign;
-	unsigned int	num;
+	int	sign;
+	int	num;
 
 	sign = 1;
 	num = 0;
@@ -27,9 +27,9 @@ int	ft_atoi(const char *str)
 			sign *= -1;
 	while (*str >= '0' && *str <= '9')
 		num = num * 10 + (*str++ - 48);
-	if (num > 2147483647 && sign == 1)
+	if ((unsigned int)num > 2147483647 && sign == 1)
 		return (-1);
-	if (num > 2147483648 && sign == -1)
+	if ((unsigned int)num > 2147483648 && sign == -1)
 		return (0);
 	return (num * sign);
 }
